Crate rusoto_elbv2

Elastic Load Balancing
A load balancer distributes incoming traffic across targets, such as your EC2 instances. This enables you to increase the availability of your application. The load balancer also monitors the health of its registered targets and ensures that it routes traffic only to healthy targets. You configure your load balancer to accept incoming traffic by specifying one or more listeners, which are configured with a protocol and port number for connections from clients to the load balancer. You configure a target group with a protocol and port number for connections from the load balancer to the targets, and with health check settings to be used when checking the health status of the targets.

Elastic Load Balancing supports the following types of load balancers: Application Load Balancers, Network Load Balancers, and Classic Load Balancers.

An Application Load Balancer makes routing and load balancing decisions at the application layer (HTTP/HTTPS). A Network Load Balancer makes routing and load balancing decisions at the transport layer (TCP/TLS). Both Application Load Balancers and Network Load Balancers can route requests to one or more ports on each EC2 instance or container instance in your virtual private cloud (VPC).

A Classic Load Balancer makes routing and load balancing decisions either at the transport layer (TCP/SSL) or the application layer (HTTP/HTTPS), and supports either EC2-Classic or a VPC. For more information, see the Elastic Load Balancing User Guide.

This reference covers the 2015-12-01 API, which supports Application Load Balancers and Network Load Balancers. The 2012-06-01 API supports Classic Load Balancers.

To get started, complete the following tasks:

Create a load balancer using CreateLoadBalancer.

Create a target group using CreateTargetGroup.

Register targets for the target group using RegisterTargets.

Create one or more listeners for your load balancer using CreateListener.

To delete a load balancer and its related resources, complete the following tasks:

Delete the load balancer using DeleteLoadBalancer.

Delete the target group using DeleteTargetGroup.

All Elastic Load Balancing operations are idempotent, which means that they complete at most one time. If you repeat an operation, it succeeds.

If you're using the service, you're probably looking for ElbClient and Elb.

Structs

Action
Information about an action.

AddListenerCertificatesInput

AddListenerCertificatesOutput

AddTagsInput

AddTagsOutput

AuthenticateCognitoActionConfig
Request parameters to use when integrating with Amazon Cognito to authenticate users.

AuthenticateOidcActionConfig
Request parameters when using an identity provider (IdP) that is compliant with OpenID Connect (OIDC) to authenticate users.

AvailabilityZone
Information about an Availability Zone.

Certificate
Information about an SSL server certificate.

Cipher
Information about a cipher used in a policy.

CreateListenerInput

CreateListenerOutput

CreateLoadBalancerInput

CreateLoadBalancerOutput

CreateRuleInput

CreateRuleOutput

CreateTargetGroupInput

CreateTargetGroupOutput

DeleteListenerInput

DeleteListenerOutput

DeleteLoadBalancerInput

DeleteLoadBalancerOutput

DeleteRuleInput

DeleteRuleOutput

DeleteTargetGroupInput

DeleteTargetGroupOutput

DeregisterTargetsInput

DeregisterTargetsOutput

DescribeAccountLimitsInput

DescribeAccountLimitsOutput

DescribeListenerCertificatesInput

DescribeListenerCertificatesOutput

DescribeListenersInput

DescribeListenersOutput

DescribeLoadBalancerAttributesInput

DescribeLoadBalancerAttributesOutput

DescribeLoadBalancersInput

DescribeLoadBalancersOutput

DescribeRulesInput

DescribeRulesOutput

DescribeSSLPoliciesInput

DescribeSSLPoliciesOutput

DescribeTagsInput

DescribeTagsOutput

DescribeTargetGroupAttributesInput

DescribeTargetGroupAttributesOutput

DescribeTargetGroupsInput

DescribeTargetGroupsOutput

DescribeTargetHealthInput

DescribeTargetHealthOutput

ElbClient
A client for the Elastic Load Balancing v2 API.

FixedResponseActionConfig
Information about an action that returns a custom HTTP response.

HostHeaderConditionConfig

HttpHeaderConditionConfig

HttpRequestMethodConditionConfig

Limit
Information about an Elastic Load Balancing resource limit for your AWS account.

Listener
Information about a listener.

LoadBalancer
Information about a load balancer.

LoadBalancerAddress
Information about a static IP address for a load balancer.

LoadBalancerAttribute
Information about a load balancer attribute.

LoadBalancerState
Information about the state of the load balancer.

Matcher
Information to use when checking for a successful response from a target.

ModifyListenerInput

ModifyListenerOutput

ModifyLoadBalancerAttributesInput

ModifyLoadBalancerAttributesOutput

ModifyRuleInput

ModifyRuleOutput

ModifyTargetGroupAttributesInput

ModifyTargetGroupAttributesOutput

ModifyTargetGroupInput

ModifyTargetGroupOutput

PathPatternConditionConfig

QueryStringConditionConfig

QueryStringKeyValuePair

RedirectActionConfig
Information about a redirect action.

A URI consists of the following components: protocol://hostname:port/path?query. You must modify at least one of the following components to avoid a redirect loop: protocol, hostname, port, or path. Any components that you do not modify retain their original values.

You can reuse URI components using the following reserved keywords:

#{protocol}

#{host}

#{port}

#{path} (the leading "/" is removed)

#{query}

For example, you can change the path to "/new/#{path}", the hostname to "example.#{host}", or the query to "#{query}&value=xyz".

RegisterTargetsInput

RegisterTargetsOutput

RemoveListenerCertificatesInput

RemoveListenerCertificatesOutput

RemoveTagsInput

RemoveTagsOutput

Rule
Information about a rule.

RuleCondition
Information about a condition for a rule.

RulePriorityPair
Information about the priorities for the rules for a listener.

SetIpAddressTypeInput

SetIpAddressTypeOutput

SetRulePrioritiesInput

SetRulePrioritiesOutput

SetSecurityGroupsInput

SetSecurityGroupsOutput

SetSubnetsInput

SetSubnetsOutput

SourceIpConditionConfig

SslPolicy
Information about a policy used for SSL negotiation.

SubnetMapping
Information about a subnet mapping.

Tag
Information about a tag.

TagDescription
The tags associated with a resource.

TargetDescription
Information about a target.

TargetGroup
Information about a target group.

TargetGroupAttribute
Information about a target group attribute.

TargetHealth
Information about the current health of a target.

TargetHealthDescription
Information about the health of a target.

Enums

AddListenerCertificatesError
Errors returned by AddListenerCertificates

AddTagsError
Errors returned by AddTags

CreateListenerError
Errors returned by CreateListener

CreateLoadBalancerError
Errors returned by CreateLoadBalancer

CreateRuleError
Errors returned by CreateRule

CreateTargetGroupError
Errors returned by CreateTargetGroup

DeleteListenerError
Errors returned by DeleteListener

DeleteLoadBalancerError
Errors returned by DeleteLoadBalancer

DeleteRuleError
Errors returned by DeleteRule

DeleteTargetGroupError
Errors returned by DeleteTargetGroup

DeregisterTargetsError
Errors returned by DeregisterTargets

DescribeAccountLimitsError
Errors returned by DescribeAccountLimits

DescribeListenerCertificatesError
Errors returned by DescribeListenerCertificates

DescribeListenersError
Errors returned by DescribeListeners

DescribeLoadBalancerAttributesError
Errors returned by DescribeLoadBalancerAttributes

DescribeLoadBalancersError
Errors returned by DescribeLoadBalancers

DescribeRulesError
Errors returned by DescribeRules

DescribeSSLPoliciesError
Errors returned by DescribeSSLPolicies

DescribeTagsError
Errors returned by DescribeTags

DescribeTargetGroupAttributesError
Errors returned by DescribeTargetGroupAttributes

DescribeTargetGroupsError
Errors returned by DescribeTargetGroups

DescribeTargetHealthError
Errors returned by DescribeTargetHealth

ModifyListenerError
Errors returned by ModifyListener

ModifyLoadBalancerAttributesError
Errors returned by ModifyLoadBalancerAttributes

ModifyRuleError
Errors returned by ModifyRule

ModifyTargetGroupAttributesError
Errors returned by ModifyTargetGroupAttributes

ModifyTargetGroupError
Errors returned by ModifyTargetGroup

RegisterTargetsError
Errors returned by RegisterTargets

RemoveListenerCertificatesError
Errors returned by RemoveListenerCertificates

RemoveTagsError
Errors returned by RemoveTags

SetIpAddressTypeError
Errors returned by SetIpAddressType

SetRulePrioritiesError
Errors returned by SetRulePriorities

SetSecurityGroupsError
Errors returned by SetSecurityGroups

SetSubnetsError
Errors returned by SetSubnets

Traits

Elb
Trait representing the capabilities of the Elastic Load Balancing v2 API. Elastic Load Balancing v2 clients implement this trait.