Enum rusoto_elbv2::ModifyListenerError

pub enum ModifyListenerError {
    CertificateNotFound(String),
    DuplicateListener(String),
    IncompatibleProtocols(String),
    InvalidConfigurationRequest(String),
    ListenerNotFound(String),
    SSLPolicyNotFound(String),
    TargetGroupAssociationLimit(String),
    TargetGroupNotFound(String),
    TooManyCertificates(String),
    TooManyListeners(String),
    TooManyRegistrationsForTargetId(String),
    TooManyTargets(String),
    UnsupportedProtocol(String),
    HttpDispatch(HttpDispatchError),
    Credentials(CredentialsError),
    Validation(String),
    Unknown(String),
}

Errors returned by ModifyListener

Variants
CertificateNotFound(String)

The specified certificate does not exist.

DuplicateListener(String)

A listener with the specified port already exists.

IncompatibleProtocols(String)

The specified configuration is not valid with this protocol.

InvalidConfigurationRequest(String)

The requested configuration is not valid.

ListenerNotFound(String)

The specified listener does not exist.

SSLPolicyNotFound(String)

The specified SSL policy does not exist.

TargetGroupAssociationLimit(String)

You've reached the limit on the number of load balancers per target group.

TargetGroupNotFound(String)

The specified target group does not exist.

TooManyCertificates(String)

You've reached the limit on the number of certificates per load balancer.

TooManyListeners(String)

You've reached the limit on the number of listeners per load balancer.

TooManyRegistrationsForTargetId(String)

You've reached the limit on the number of times a target can be registered with a load balancer.

TooManyTargets(String)

You've reached the limit on the number of targets.

UnsupportedProtocol(String)

The specified protocol is not supported.

HttpDispatch(HttpDispatchError)

An error occurred dispatching the HTTP request

Credentials(CredentialsError)

An error was encountered with AWS credentials.

Validation(String)

A validation error occurred. Details from AWS are provided.

Unknown(String)

An unknown error occurred. The raw HTTP response is provided.

Methods

impl ModifyListenerError

pub fn from_body(body: &str) -> ModifyListenerError