Struct rusoto_elbv2::Action

pub struct Action {
    pub authenticate_cognito_config: Option<AuthenticateCognitoActionConfig>,
    pub authenticate_oidc_config: Option<AuthenticateOidcActionConfig>,
    pub fixed_response_config: Option<FixedResponseActionConfig>,
    pub order: Option<i64>,
    pub redirect_config: Option<RedirectActionConfig>,
    pub target_group_arn: Option<String>,
    pub type_: String,
}

Information about an action.

Fields
authenticate_cognito_config: Option<AuthenticateCognitoActionConfig>

[HTTPS listener] Information for using Amazon Cognito to authenticate users. Specify only when Type is authenticate-cognito.

authenticate_oidc_config: Option<AuthenticateOidcActionConfig>

[HTTPS listener] Information about an identity provider that is compliant with OpenID Connect (OIDC). Specify only when Type is authenticate-oidc.

fixed_response_config: Option<FixedResponseActionConfig>

[Application Load Balancer] Information for creating an action that returns a custom HTTP response. Specify only when Type is fixed-response.

order: Option<i64>

The order for the action. This value is required for rules with multiple actions. The action with the lowest value for order is performed first. The final action to be performed must be a forward or a fixed-response action.

redirect_config: Option<RedirectActionConfig>

[Application Load Balancer] Information for creating a redirect action. Specify only when Type is redirect.

target_group_arn: Option<String>

The Amazon Resource Name (ARN) of the target group. Specify only when Type is forward.

type_: String

The type of action. Each rule must include exactly one of the following types of actions: forward, fixed-response, or redirect.