Struct rusoto_elbv2::CreateTargetGroupInput

pub struct CreateTargetGroupInput {
    pub health_check_interval_seconds: Option<i64>,
    pub health_check_path: Option<String>,
    pub health_check_port: Option<String>,
    pub health_check_protocol: Option<String>,
    pub health_check_timeout_seconds: Option<i64>,
    pub healthy_threshold_count: Option<i64>,
    pub matcher: Option<Matcher>,
    pub name: String,
    pub port: i64,
    pub protocol: String,
    pub target_type: Option<String>,
    pub unhealthy_threshold_count: Option<i64>,
    pub vpc_id: String,
}

Fields
health_check_interval_seconds: Option<i64>

The approximate amount of time, in seconds, between health checks of an individual target. For Application Load Balancers, the range is 5 to 300 seconds. For Network Load Balancers, the supported values are 10 or 30 seconds. The default is 30 seconds.

health_check_path: Option<String>

[HTTP/HTTPS health checks] The ping path that is the destination on the targets for health checks. The default is /.

health_check_port: Option<String>

The port the load balancer uses when performing health checks on targets. The default is traffic-port, which is the port on which each target receives traffic from the load balancer.

health_check_protocol: Option<String>

The protocol the load balancer uses when performing health checks on targets. The TCP protocol is supported only if the protocol of the target group is TCP. For Application Load Balancers, the default is HTTP. For Network Load Balancers, the default is TCP.

health_check_timeout_seconds: Option<i64>

The amount of time, in seconds, during which no response from a target means a failed health check. For Application Load Balancers, the range is 2 to 60 seconds and the default is 5 seconds. For Network Load Balancers, this is 10 seconds for TCP and HTTPS health checks and 6 seconds for HTTP health checks.

healthy_threshold_count: Option<i64>

The number of consecutive health checks successes required before considering an unhealthy target healthy. For Application Load Balancers, the default is 5. For Network Load Balancers, the default is 3.

matcher: Option<Matcher>

[HTTP/HTTPS health checks] The HTTP codes to use when checking for a successful response from a target.

name: String

The name of the target group.

This name must be unique per region per account, can have a maximum of 32 characters, must contain only alphanumeric characters or hyphens, and must not begin or end with a hyphen.

port: i64

The port on which the targets receive traffic. This port is used unless you specify a port override when registering the target.

protocol: String

The protocol to use for routing traffic to the targets. For Application Load Balancers, the supported protocols are HTTP and HTTPS. For Network Load Balancers, the supported protocol is TCP.

target_type: Option<String>

The type of target that you must specify when registering targets with this target group. The possible values are instance (targets are specified by instance ID) or ip (targets are specified by IP address). The default is instance. Note that you can't specify targets for a target group using both instance IDs and IP addresses.

If the target type is ip, specify IP addresses from the subnets of the virtual private cloud (VPC) for the target group, the RFC 1918 range (10.0.0.0/8, 172.16.0.0/12, and 192.168.0.0/16), and the RFC 6598 range (100.64.0.0/10). You can't specify publicly routable IP addresses.

unhealthy_threshold_count: Option<i64>

The number of consecutive health check failures required before considering a target unhealthy. For Application Load Balancers, the default is 2. For Network Load Balancers, this value must be the same as the healthy threshold count.

vpc_id: String

The identifier of the virtual private cloud (VPC).