Struct rusoto_elbv2::ModifyTargetGroupInput

pub struct ModifyTargetGroupInput {
    pub health_check_interval_seconds: Option<i64>,
    pub health_check_path: Option<String>,
    pub health_check_port: Option<String>,
    pub health_check_protocol: Option<String>,
    pub health_check_timeout_seconds: Option<i64>,
    pub healthy_threshold_count: Option<i64>,
    pub matcher: Option<Matcher>,
    pub target_group_arn: String,
    pub unhealthy_threshold_count: Option<i64>,
}

Fields
health_check_interval_seconds: Option<i64>

The approximate amount of time, in seconds, between health checks of an individual target.

health_check_path: Option<String>

The ping path that is the destination for the health check request.

health_check_port: Option<String>

The port to use to connect with the target.

health_check_protocol: Option<String>

The protocol to use to connect with the target.

health_check_timeout_seconds: Option<i64>

The amount of time, in seconds, during which no response means a failed health check.

healthy_threshold_count: Option<i64>

The number of consecutive health checks successes required before considering an unhealthy target healthy.

matcher: Option<Matcher>

The HTTP codes to use when checking for a successful response from a target.

target_group_arn: String

The Amazon Resource Name (ARN) of the target group.

unhealthy_threshold_count: Option<i64>

The number of consecutive health check failures required before considering the target unhealthy.