Trait rusoto_elbv2::Elb

pub trait Elb {
    fn add_listener_certificates(
        &self, 
        input: AddListenerCertificatesInput
    ) -> RusotoFuture<AddListenerCertificatesOutput, AddListenerCertificatesError>;

    fn add_tags(
        &self, 
        input: AddTagsInput
    ) -> RusotoFuture<AddTagsOutput, AddTagsError>;

    fn create_listener(
        &self, 
        input: CreateListenerInput
    ) -> RusotoFuture<CreateListenerOutput, CreateListenerError>;

    fn create_load_balancer(
        &self, 
        input: CreateLoadBalancerInput
    ) -> RusotoFuture<CreateLoadBalancerOutput, CreateLoadBalancerError>;

    fn create_rule(
        &self, 
        input: CreateRuleInput
    ) -> RusotoFuture<CreateRuleOutput, CreateRuleError>;

    fn create_target_group(
        &self, 
        input: CreateTargetGroupInput
    ) -> RusotoFuture<CreateTargetGroupOutput, CreateTargetGroupError>;

    fn delete_listener(
        &self, 
        input: DeleteListenerInput
    ) -> RusotoFuture<DeleteListenerOutput, DeleteListenerError>;

    fn delete_load_balancer(
        &self, 
        input: DeleteLoadBalancerInput
    ) -> RusotoFuture<DeleteLoadBalancerOutput, DeleteLoadBalancerError>;

    fn delete_rule(
        &self, 
        input: DeleteRuleInput
    ) -> RusotoFuture<DeleteRuleOutput, DeleteRuleError>;

    fn delete_target_group(
        &self, 
        input: DeleteTargetGroupInput
    ) -> RusotoFuture<DeleteTargetGroupOutput, DeleteTargetGroupError>;

    fn deregister_targets(
        &self, 
        input: DeregisterTargetsInput
    ) -> RusotoFuture<DeregisterTargetsOutput, DeregisterTargetsError>;

    fn describe_account_limits(
        &self, 
        input: DescribeAccountLimitsInput
    ) -> RusotoFuture<DescribeAccountLimitsOutput, DescribeAccountLimitsError>;

    fn describe_listener_certificates(
        &self, 
        input: DescribeListenerCertificatesInput
    ) -> RusotoFuture<DescribeListenerCertificatesOutput, DescribeListenerCertificatesError>;

    fn describe_listeners(
        &self, 
        input: DescribeListenersInput
    ) -> RusotoFuture<DescribeListenersOutput, DescribeListenersError>;

    fn describe_load_balancer_attributes(
        &self, 
        input: DescribeLoadBalancerAttributesInput
    ) -> RusotoFuture<DescribeLoadBalancerAttributesOutput, DescribeLoadBalancerAttributesError>;

    fn describe_load_balancers(
        &self, 
        input: DescribeLoadBalancersInput
    ) -> RusotoFuture<DescribeLoadBalancersOutput, DescribeLoadBalancersError>;

    fn describe_rules(
        &self, 
        input: DescribeRulesInput
    ) -> RusotoFuture<DescribeRulesOutput, DescribeRulesError>;

    fn describe_ssl_policies(
        &self, 
        input: DescribeSSLPoliciesInput
    ) -> RusotoFuture<DescribeSSLPoliciesOutput, DescribeSSLPoliciesError>;

    fn describe_tags(
        &self, 
        input: DescribeTagsInput
    ) -> RusotoFuture<DescribeTagsOutput, DescribeTagsError>;

    fn describe_target_group_attributes(
        &self, 
        input: DescribeTargetGroupAttributesInput
    ) -> RusotoFuture<DescribeTargetGroupAttributesOutput, DescribeTargetGroupAttributesError>;

    fn describe_target_groups(
        &self, 
        input: DescribeTargetGroupsInput
    ) -> RusotoFuture<DescribeTargetGroupsOutput, DescribeTargetGroupsError>;

    fn describe_target_health(
        &self, 
        input: DescribeTargetHealthInput
    ) -> RusotoFuture<DescribeTargetHealthOutput, DescribeTargetHealthError>;

    fn modify_listener(
        &self, 
        input: ModifyListenerInput
    ) -> RusotoFuture<ModifyListenerOutput, ModifyListenerError>;

    fn modify_load_balancer_attributes(
        &self, 
        input: ModifyLoadBalancerAttributesInput
    ) -> RusotoFuture<ModifyLoadBalancerAttributesOutput, ModifyLoadBalancerAttributesError>;

    fn modify_rule(
        &self, 
        input: ModifyRuleInput
    ) -> RusotoFuture<ModifyRuleOutput, ModifyRuleError>;

    fn modify_target_group(
        &self, 
        input: ModifyTargetGroupInput
    ) -> RusotoFuture<ModifyTargetGroupOutput, ModifyTargetGroupError>;

    fn modify_target_group_attributes(
        &self, 
        input: ModifyTargetGroupAttributesInput
    ) -> RusotoFuture<ModifyTargetGroupAttributesOutput, ModifyTargetGroupAttributesError>;

    fn register_targets(
        &self, 
        input: RegisterTargetsInput
    ) -> RusotoFuture<RegisterTargetsOutput, RegisterTargetsError>;

    fn remove_listener_certificates(
        &self, 
        input: RemoveListenerCertificatesInput
    ) -> RusotoFuture<RemoveListenerCertificatesOutput, RemoveListenerCertificatesError>;

    fn remove_tags(
        &self, 
        input: RemoveTagsInput
    ) -> RusotoFuture<RemoveTagsOutput, RemoveTagsError>;

    fn set_ip_address_type(
        &self, 
        input: SetIpAddressTypeInput
    ) -> RusotoFuture<SetIpAddressTypeOutput, SetIpAddressTypeError>;

    fn set_rule_priorities(
        &self, 
        input: SetRulePrioritiesInput
    ) -> RusotoFuture<SetRulePrioritiesOutput, SetRulePrioritiesError>;

    fn set_security_groups(
        &self, 
        input: SetSecurityGroupsInput
    ) -> RusotoFuture<SetSecurityGroupsOutput, SetSecurityGroupsError>;

    fn set_subnets(
        &self, 
        input: SetSubnetsInput
    ) -> RusotoFuture<SetSubnetsOutput, SetSubnetsError>;
}

Trait representing the capabilities of the Elastic Load Balancing v2 API. Elastic Load Balancing v2 clients implement this trait.

Required methods

fn add_listener_certificates(
&self,
input: AddListenerCertificatesInput
) -> RusotoFuture<AddListenerCertificatesOutput, AddListenerCertificatesError>

Adds the specified certificate to the specified HTTPS listener.

If the certificate was already added, the call is successful but the certificate is not added again.

To list the certificates for your listener, use DescribeListenerCertificates. To remove certificates from your listener, use RemoveListenerCertificates. To specify the default SSL server certificate, use ModifyListener.

fn add_tags(
&self,
input: AddTagsInput
) -> RusotoFuture<AddTagsOutput, AddTagsError>

Adds the specified tags to the specified Elastic Load Balancing resource. You can tag your Application Load Balancers, Network Load Balancers, and your target groups.

Each tag consists of a key and an optional value. If a resource already has a tag with the same key, AddTags updates its value.

To list the current tags for your resources, use DescribeTags. To remove tags from your resources, use RemoveTags.

fn create_listener(
&self,
input: CreateListenerInput
) -> RusotoFuture<CreateListenerOutput, CreateListenerError>

Creates a listener for the specified Application Load Balancer or Network Load Balancer.

To update a listener, use ModifyListener. When you are finished with a listener, you can delete it using DeleteListener. If you are finished with both the listener and the load balancer, you can delete them both using DeleteLoadBalancer.

This operation is idempotent, which means that it completes at most one time. If you attempt to create multiple listeners with the same settings, each call succeeds.

For more information, see Listeners for Your Application Load Balancers in the Application Load Balancers Guide and Listeners for Your Network Load Balancers in the Network Load Balancers Guide.

fn create_load_balancer(
&self,
input: CreateLoadBalancerInput
) -> RusotoFuture<CreateLoadBalancerOutput, CreateLoadBalancerError>

Creates an Application Load Balancer or a Network Load Balancer.

When you create a load balancer, you can specify security groups, public subnets, IP address type, and tags. Otherwise, you could do so later using SetSecurityGroups, SetSubnets, SetIpAddressType, and AddTags.

To create listeners for your load balancer, use CreateListener. To describe your current load balancers, see DescribeLoadBalancers. When you are finished with a load balancer, you can delete it using DeleteLoadBalancer.

For limit information, see Limits for Your Application Load Balancer in the Application Load Balancers Guide and Limits for Your Network Load Balancer in the Network Load Balancers Guide.

This operation is idempotent, which means that it completes at most one time. If you attempt to create multiple load balancers with the same settings, each call succeeds.

For more information, see Application Load Balancers in the Application Load Balancers Guide and Network Load Balancers in the Network Load Balancers Guide.

fn create_rule(
&self,
input: CreateRuleInput
) -> RusotoFuture<CreateRuleOutput, CreateRuleError>

Creates a rule for the specified listener. The listener must be associated with an Application Load Balancer.

Rules are evaluated in priority order, from the lowest value to the highest value. When the conditions for a rule are met, its actions are performed. If the conditions for no rules are met, the actions for the default rule are performed. For more information, see Listener Rules in the Application Load Balancers Guide.

To view your current rules, use DescribeRules. To update a rule, use ModifyRule. To set the priorities of your rules, use SetRulePriorities. To delete a rule, use DeleteRule.

fn create_target_group(
&self,
input: CreateTargetGroupInput
) -> RusotoFuture<CreateTargetGroupOutput, CreateTargetGroupError>

Creates a target group.

To register targets with the target group, use RegisterTargets. To update the health check settings for the target group, use ModifyTargetGroup. To monitor the health of targets in the target group, use DescribeTargetHealth.

To route traffic to the targets in a target group, specify the target group in an action using CreateListener or CreateRule.

To delete a target group, use DeleteTargetGroup.

This operation is idempotent, which means that it completes at most one time. If you attempt to create multiple target groups with the same settings, each call succeeds.

For more information, see Target Groups for Your Application Load Balancers in the Application Load Balancers Guide or Target Groups for Your Network Load Balancers in the Network Load Balancers Guide.

fn delete_listener(
&self,
input: DeleteListenerInput
) -> RusotoFuture<DeleteListenerOutput, DeleteListenerError>

Deletes the specified listener.

Alternatively, your listener is deleted when you delete the load balancer to which it is attached, using DeleteLoadBalancer.

fn delete_load_balancer(
&self,
input: DeleteLoadBalancerInput
) -> RusotoFuture<DeleteLoadBalancerOutput, DeleteLoadBalancerError>

Deletes the specified Application Load Balancer or Network Load Balancer and its attached listeners.

You can't delete a load balancer if deletion protection is enabled. If the load balancer does not exist or has already been deleted, the call succeeds.

Deleting a load balancer does not affect its registered targets. For example, your EC2 instances continue to run and are still registered to their target groups. If you no longer need these EC2 instances, you can stop or terminate them.

fn delete_rule(
&self,
input: DeleteRuleInput
) -> RusotoFuture<DeleteRuleOutput, DeleteRuleError>

Deletes the specified rule.

fn delete_target_group(
&self,
input: DeleteTargetGroupInput
) -> RusotoFuture<DeleteTargetGroupOutput, DeleteTargetGroupError>

Deletes the specified target group.

You can delete a target group if it is not referenced by any actions. Deleting a target group also deletes any associated health checks.

fn deregister_targets(
&self,
input: DeregisterTargetsInput
) -> RusotoFuture<DeregisterTargetsOutput, DeregisterTargetsError>

Deregisters the specified targets from the specified target group. After the targets are deregistered, they no longer receive traffic from the load balancer.

fn describe_account_limits(
&self,
input: DescribeAccountLimitsInput
) -> RusotoFuture<DescribeAccountLimitsOutput, DescribeAccountLimitsError>

Describes the current Elastic Load Balancing resource limits for your AWS account.

For more information, see Limits for Your Application Load Balancers in the Application Load Balancer Guide or Limits for Your Network Load Balancers in the Network Load Balancers Guide.

fn describe_listener_certificates(
&self,
input: DescribeListenerCertificatesInput
) -> RusotoFuture<DescribeListenerCertificatesOutput, DescribeListenerCertificatesError>

Describes the certificates for the specified HTTPS listener.

fn describe_listeners(
&self,
input: DescribeListenersInput
) -> RusotoFuture<DescribeListenersOutput, DescribeListenersError>

Describes the specified listeners or the listeners for the specified Application Load Balancer or Network Load Balancer. You must specify either a load balancer or one or more listeners.

fn describe_load_balancer_attributes(
&self,
input: DescribeLoadBalancerAttributesInput
) -> RusotoFuture<DescribeLoadBalancerAttributesOutput, DescribeLoadBalancerAttributesError>

Describes the attributes for the specified Application Load Balancer or Network Load Balancer.

For more information, see Load Balancer Attributes in the Application Load Balancers Guide or Load Balancer Attributes in the Network Load Balancers Guide.

fn describe_load_balancers(
&self,
input: DescribeLoadBalancersInput
) -> RusotoFuture<DescribeLoadBalancersOutput, DescribeLoadBalancersError>

Describes the specified load balancers or all of your load balancers.

To describe the listeners for a load balancer, use DescribeListeners. To describe the attributes for a load balancer, use DescribeLoadBalancerAttributes.

fn describe_rules(
&self,
input: DescribeRulesInput
) -> RusotoFuture<DescribeRulesOutput, DescribeRulesError>

Describes the specified rules or the rules for the specified listener. You must specify either a listener or one or more rules.

fn describe_ssl_policies(
&self,
input: DescribeSSLPoliciesInput
) -> RusotoFuture<DescribeSSLPoliciesOutput, DescribeSSLPoliciesError>

Describes the specified policies or all policies used for SSL negotiation.

For more information, see Security Policies in the Application Load Balancers Guide.

fn describe_tags(
&self,
input: DescribeTagsInput
) -> RusotoFuture<DescribeTagsOutput, DescribeTagsError>

Describes the tags for the specified resources. You can describe the tags for one or more Application Load Balancers, Network Load Balancers, and target groups.

fn describe_target_group_attributes(
&self,
input: DescribeTargetGroupAttributesInput
) -> RusotoFuture<DescribeTargetGroupAttributesOutput, DescribeTargetGroupAttributesError>

Describes the attributes for the specified target group.

For more information, see Target Group Attributes in the Application Load Balancers Guide or Target Group Attributes in the Network Load Balancers Guide.

fn describe_target_groups(
&self,
input: DescribeTargetGroupsInput
) -> RusotoFuture<DescribeTargetGroupsOutput, DescribeTargetGroupsError>

Describes the specified target groups or all of your target groups. By default, all target groups are described. Alternatively, you can specify one of the following to filter the results: the ARN of the load balancer, the names of one or more target groups, or the ARNs of one or more target groups.

To describe the targets for a target group, use DescribeTargetHealth. To describe the attributes of a target group, use DescribeTargetGroupAttributes.

fn describe_target_health(
&self,
input: DescribeTargetHealthInput
) -> RusotoFuture<DescribeTargetHealthOutput, DescribeTargetHealthError>

Describes the health of the specified targets or all of your targets.

fn modify_listener(
&self,
input: ModifyListenerInput
) -> RusotoFuture<ModifyListenerOutput, ModifyListenerError>

Modifies the specified properties of the specified listener.

Any properties that you do not specify retain their current values. However, changing the protocol from HTTPS to HTTP, or from TLS to TCP, removes the security policy and server certificate properties. If you change the protocol from HTTP to HTTPS, or from TCP to TLS, you must add the security policy and server certificate properties.

fn modify_load_balancer_attributes(
&self,
input: ModifyLoadBalancerAttributesInput
) -> RusotoFuture<ModifyLoadBalancerAttributesOutput, ModifyLoadBalancerAttributesError>

Modifies the specified attributes of the specified Application Load Balancer or Network Load Balancer.

If any of the specified attributes can't be modified as requested, the call fails. Any existing attributes that you do not modify retain their current values.

fn modify_rule(
&self,
input: ModifyRuleInput
) -> RusotoFuture<ModifyRuleOutput, ModifyRuleError>

Modifies the specified rule.

Any existing properties that you do not modify retain their current values.

To modify the actions for the default rule, use ModifyListener.

fn modify_target_group(
&self,
input: ModifyTargetGroupInput
) -> RusotoFuture<ModifyTargetGroupOutput, ModifyTargetGroupError>

Modifies the health checks used when evaluating the health state of the targets in the specified target group.

To monitor the health of the targets, use DescribeTargetHealth.

fn modify_target_group_attributes(
&self,
input: ModifyTargetGroupAttributesInput
) -> RusotoFuture<ModifyTargetGroupAttributesOutput, ModifyTargetGroupAttributesError>

Modifies the specified attributes of the specified target group.

fn register_targets(
&self,
input: RegisterTargetsInput
) -> RusotoFuture<RegisterTargetsOutput, RegisterTargetsError>

Registers the specified targets with the specified target group.

If the target is an EC2 instance, it must be in the running state when you register it.

By default, the load balancer routes requests to registered targets using the protocol and port for the target group. Alternatively, you can override the port for a target when you register it. You can register each EC2 instance or IP address with the same target group multiple times using different ports.

With a Network Load Balancer, you cannot register instances by instance ID if they have the following instance types: C1, CC1, CC2, CG1, CG2, CR1, CS1, G1, G2, HI1, HS1, M1, M2, M3, and T1. You can register instances of these types by IP address.

To remove a target from a target group, use DeregisterTargets.

fn remove_listener_certificates(
&self,
input: RemoveListenerCertificatesInput
) -> RusotoFuture<RemoveListenerCertificatesOutput, RemoveListenerCertificatesError>

Removes the specified certificate from the specified HTTPS listener.

You can't remove the default certificate for a listener. To replace the default certificate, call ModifyListener.

To list the certificates for your listener, use DescribeListenerCertificates.

fn remove_tags(
&self,
input: RemoveTagsInput
) -> RusotoFuture<RemoveTagsOutput, RemoveTagsError>

Removes the specified tags from the specified Elastic Load Balancing resource.

To list the current tags for your resources, use DescribeTags.

fn set_ip_address_type(
&self,
input: SetIpAddressTypeInput
) -> RusotoFuture<SetIpAddressTypeOutput, SetIpAddressTypeError>

Sets the type of IP addresses used by the subnets of the specified Application Load Balancer or Network Load Balancer.

Network Load Balancers must use ipv4.

fn set_rule_priorities(
&self,
input: SetRulePrioritiesInput
) -> RusotoFuture<SetRulePrioritiesOutput, SetRulePrioritiesError>

Sets the priorities of the specified rules.

You can reorder the rules as long as there are no priority conflicts in the new order. Any existing rules that you do not specify retain their current priority.

fn set_security_groups(
&self,
input: SetSecurityGroupsInput
) -> RusotoFuture<SetSecurityGroupsOutput, SetSecurityGroupsError>

Associates the specified security groups with the specified Application Load Balancer. The specified security groups override the previously associated security groups.

You can't specify a security group for a Network Load Balancer.

fn set_subnets(
&self,
input: SetSubnetsInput
) -> RusotoFuture<SetSubnetsOutput, SetSubnetsError>

Enables the Availability Zone for the specified public subnets for the specified Application Load Balancer. The specified subnets replace the previously enabled subnets.

You can't change the subnets for a Network Load Balancer.

Loading content...