Crate rusoto_elbv2

Elastic Load Balancing

If you're using the service, you're probably looking for ElbClient and Elb.

Structs

Action
Information about an action.

AddTagsInput

AddTagsOutput

AvailabilityZone
Information about an Availability Zone.

Certificate
Information about an SSL server certificate deployed on a load balancer.

Cipher
Information about a cipher used in a policy.

CreateListenerInput

CreateListenerOutput

CreateLoadBalancerInput

CreateLoadBalancerOutput

CreateRuleInput

CreateRuleOutput

CreateTargetGroupInput

CreateTargetGroupOutput

DeleteListenerInput

DeleteListenerOutput

DeleteLoadBalancerInput

DeleteLoadBalancerOutput

DeleteRuleInput

DeleteRuleOutput

DeleteTargetGroupInput

DeleteTargetGroupOutput

DeregisterTargetsInput

DeregisterTargetsOutput

DescribeAccountLimitsInput

DescribeAccountLimitsOutput

DescribeListenersInput

DescribeListenersOutput

DescribeLoadBalancerAttributesInput

DescribeLoadBalancerAttributesOutput

DescribeLoadBalancersInput

DescribeLoadBalancersOutput

DescribeRulesInput

DescribeRulesOutput

DescribeSSLPoliciesInput

DescribeSSLPoliciesOutput

DescribeTagsInput

DescribeTagsOutput

DescribeTargetGroupAttributesInput

DescribeTargetGroupAttributesOutput

DescribeTargetGroupsInput

DescribeTargetGroupsOutput

DescribeTargetHealthInput

DescribeTargetHealthOutput

ElbClient
A client for the Elastic Load Balancing v2 API.

Limit
Information about an Elastic Load Balancing resource limit for your AWS account.

Listener
Information about a listener.

LoadBalancer
Information about a load balancer.

LoadBalancerAttribute
Information about a load balancer attribute.

LoadBalancerState
Information about the state of the load balancer.

Matcher
Information to use when checking for a successful response from a target.

ModifyListenerInput

ModifyListenerOutput

ModifyLoadBalancerAttributesInput

ModifyLoadBalancerAttributesOutput

ModifyRuleInput

ModifyRuleOutput

ModifyTargetGroupAttributesInput

ModifyTargetGroupAttributesOutput

ModifyTargetGroupInput

ModifyTargetGroupOutput

RegisterTargetsInput

RegisterTargetsOutput

RemoveTagsInput

RemoveTagsOutput

Rule
Information about a rule.

RuleCondition
Information about a condition for a rule.

RulePriorityPair
Information about the priorities for the rules for a listener.

SetIpAddressTypeInput

SetIpAddressTypeOutput

SetRulePrioritiesInput

SetRulePrioritiesOutput

SetSecurityGroupsInput

SetSecurityGroupsOutput

SetSubnetsInput

SetSubnetsOutput

SslPolicy
Information about a policy used for SSL negotiation.

Tag
Information about a tag.

TagDescription
The tags associated with a resource.

TargetDescription
Information about a target.

TargetGroup
Information about a target group.

TargetGroupAttribute
Information about a target group attribute.

TargetHealth
Information about the current health of a target.

TargetHealthDescription
Information about the health of a target.

Enums

AddTagsError
Errors returned by AddTags

CreateListenerError
Errors returned by CreateListener

CreateLoadBalancerError
Errors returned by CreateLoadBalancer

CreateRuleError
Errors returned by CreateRule

CreateTargetGroupError
Errors returned by CreateTargetGroup

DeleteListenerError
Errors returned by DeleteListener

DeleteLoadBalancerError
Errors returned by DeleteLoadBalancer

DeleteRuleError
Errors returned by DeleteRule

DeleteTargetGroupError
Errors returned by DeleteTargetGroup

DeregisterTargetsError
Errors returned by DeregisterTargets

DescribeAccountLimitsError
Errors returned by DescribeAccountLimits

DescribeListenersError
Errors returned by DescribeListeners

DescribeLoadBalancerAttributesError
Errors returned by DescribeLoadBalancerAttributes

DescribeLoadBalancersError
Errors returned by DescribeLoadBalancers

DescribeRulesError
Errors returned by DescribeRules

DescribeSSLPoliciesError
Errors returned by DescribeSSLPolicies

DescribeTagsError
Errors returned by DescribeTags

DescribeTargetGroupAttributesError
Errors returned by DescribeTargetGroupAttributes

DescribeTargetGroupsError
Errors returned by DescribeTargetGroups

DescribeTargetHealthError
Errors returned by DescribeTargetHealth

ModifyListenerError
Errors returned by ModifyListener

ModifyLoadBalancerAttributesError
Errors returned by ModifyLoadBalancerAttributes

ModifyRuleError
Errors returned by ModifyRule

ModifyTargetGroupAttributesError
Errors returned by ModifyTargetGroupAttributes

ModifyTargetGroupError
Errors returned by ModifyTargetGroup

RegisterTargetsError
Errors returned by RegisterTargets

RemoveTagsError
Errors returned by RemoveTags

SetIpAddressTypeError
Errors returned by SetIpAddressType

SetRulePrioritiesError
Errors returned by SetRulePriorities

SetSecurityGroupsError
Errors returned by SetSecurityGroups

SetSubnetsError
Errors returned by SetSubnets

Traits

Elb
Trait representing the capabilities of the Elastic Load Balancing v2 API. Elastic Load Balancing v2 clients implement this trait.

Type Definitions

ActionTypeEnum

Actions

AvailabilityZones

CanonicalHostedZoneId

CertificateArn

CertificateList

CipherName

CipherPriority

Ciphers

ConditionFieldName

CreatedTime

DNSName

Description

HealthCheckIntervalSeconds

HealthCheckPort

HealthCheckThresholdCount

HealthCheckTimeoutSeconds

HttpCode

IpAddressType

IsDefault

Limits

ListOfString

ListenerArn

ListenerArns

Listeners

LoadBalancerArn

LoadBalancerArns

LoadBalancerAttributeKey

LoadBalancerAttributeValue

LoadBalancerAttributes

LoadBalancerName

LoadBalancerNames

LoadBalancerSchemeEnum

LoadBalancerStateEnum

LoadBalancerTypeEnum

LoadBalancers

Marker

Max

Name

PageSize

Path

Port

ProtocolEnum

ResourceArn

ResourceArns

RuleArn

RuleArns

RuleConditionList

RulePriority

RulePriorityList

Rules

SecurityGroupId

SecurityGroups

SslPolicies

SslPolicyName

SslPolicyNames

SslProtocol

SslProtocols

StateReason

StringValue

SubnetId

Subnets

TagDescriptions

TagKey

TagKeys

TagList

TagValue

TargetDescriptions

TargetGroupArn

TargetGroupArns

TargetGroupAttributeKey

TargetGroupAttributeValue

TargetGroupAttributes

TargetGroupName

TargetGroupNames

TargetGroups

TargetHealthDescriptions

TargetHealthReasonEnum

TargetHealthStateEnum

TargetId

VpcId

ZoneName